Add boot time allocator basis.
This patch adds memblock interface for allocation and deallocation.
We maintain two arrays: one contains entries for all memory
regions (available or not), the other one contains reserved memory
blocks. To allocate memory we lookup entry in the first array and
allocate part of appropriate block such that it is not overlaped
by any entry in the second array, after that we add allocated region
in the second array so that no one can allocate it again.